+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.7.0](https://github.com/dldevinc/paper-rq/tree/v0.7.0) - 2023-08-21
+
+### Features
+
+-   Added support for Python 3.11.
+-   Added support for Django 4.2.
+-   Droped support for Django 2.2, 3.0, 3.1.
+-   Droped support for Python 3.6.
+-   Added support for `rq` >= 1.12.0.
+-   Added `job` decorator.
+-   Added `started_at` field to admin interface.
+
 ## [0.6.1](https://github.com/dldevinc/paper-rq/tree/v0.6.1) - 2022-11-18
 
 ### Features